+++ .eslintrc.js
@@ -22,6 +22,7 @@ module.exports = {
     },
     rules: {
       'jsx-a11y/label-has-for': 'off',
+      'jsx-a11y/no-noninteractive-tabindex': 'off',
     },
   }, {
     files: [