+++ src/components/custom-field-input-text/custom-field-input-text.jsx
@@ -1,63 +1,31 @@
 import PropTypes from 'prop-types';
-import React, { forwardRef, useImperativeHandle, useRef } from 'react';
-import useValidation from '../../hooks/use-validation.jsx';
-import AbstractCustomField from '../__internal__/abstract-custom-field/abstract-custom-field.jsx';
+import React, {
+  forwardRef,
+} from 'react';
+import Input from '../input/input.jsx';
 
 const CustomFieldInputText = forwardRef(function CustomFieldInputText(props, ref) {
-  const inputRef = useRef(null);
-
-  const validationMessage = useValidation(props.readOnly, props.errorText, inputRef);
-
-  useImperativeHandle(ref, () => ({
-    get dirty() {
-      const providedValue = props.defaultValue || '';
-      return providedValue !== this.value;
-    },
-    id: props.id,
-    name: props.name,
-    get value() {
-      return inputRef.current.value;
-    },
-  }));
-
   return (
-    <AbstractCustomField
-      ariaProps={props.ariaProps}
+    <Input
       className={props.className}
-      defaultValue={props.defaultValue}
-      disabled={props.disabled}
-      errorText={validationMessage}
       id={props.id}
-      inputRef={inputRef}
       label={props.label}
       maxLength={props.maxLength}
       name={props.name}
-      onBlur={props.onBlur}
-      onChange={props.onChange}
-      onClick={props.onClick}
       onFocus={props.onFocus}
-      onKeyDown={props.onKeyDown}
-      onKeyUp={props.onKeyUp}
       placeholder={props.placeholder}
       readOnly={props.readOnly}
+      ref={ref}
       required={props.required}
       type="text"
+      validationMessage={props.errorText}
       value={props.value}
     />
   );
 });
 
 CustomFieldInputText.propTypes = {
-  ariaProps: PropTypes.shape({
-    autocomplete: PropTypes.string,
-    haspopup: PropTypes.string,
-  }),
   className: PropTypes.string,
-  defaultValue: PropTypes.oneOfType([
-    PropTypes.number,
-    PropTypes.string,
-  ]),
-  disabled: PropTypes.bool,
   errorText: PropTypes.string,
   id: PropTypes.string.isRequired,
   label: PropTypes.string.isRequired,
@@ -66,35 +34,19 @@ CustomFieldInputText.propTypes = {
     PropTypes.string,
   ]),
   name: PropTypes.string,
-  onBlur: PropTypes.func,
-  onChange: PropTypes.func,
-  onClick: PropTypes.func,
   onFocus: PropTypes.func,
-  onKeyDown: PropTypes.func,
-  onKeyUp: PropTypes.func,
   placeholder: PropTypes.string,
   readOnly: PropTypes.bool,
   required: PropTypes.bool,
-  value: PropTypes.oneOfType([
-    PropTypes.number,
-    PropTypes.string,
-  ]),
+  value: PropTypes.string,
 };
 
 CustomFieldInputText.defaultProps = {
-  ariaProps: {},
   className: undefined,
-  defaultValue: undefined,
-  disabled: false,
   errorText: '',
   maxLength: undefined,
   name: undefined,
-  onBlur: () => {},
-  onChange: () => {},
-  onClick: () => {},
   onFocus: () => {},
-  onKeyDown: () => {},
-  onKeyUp: () => {},
   placeholder: undefined,
   readOnly: false,
   required: false,