+++ src/components/section-row/section-row.test.jsx
@@ -1,15 +1,11 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 import { render, screen } from '@testing-library/react';
 import SectionRow from './section-row.jsx';
 
 describe('Section Row', () => {
   it('renders a section-row component', () => {
-    const tree = renderer
-      .create((
-        <SectionRow />
-      )).toJSON();
-    expect(tree).toMatchSnapshot();
+    render(<SectionRow />);
+    expect(document.body).toMatchSnapshot();
   });
 
   it('can have child elements', () => {